Central/requirements.txt
flask==3.0.2
cassandra-driver==3.29.1
python-dotenv==1.0.1